Add koalixcrm application files and update settings
The koalixcrm application files have been added, along with the configuration settings for  production. The settings have been tailored to cover Django base settings and customized for the specific requirements of the koalixcrm project. Additionally, the instructions and data for the Docker container setup have also been updated for production deployment.